Rename Entries.py, adjust tests to fix code changes
