+++ tests/test_volumes.py
@@ -2,7 +2,6 @@ import pytest
 
 from Simyan.exceptions import APIError
 
-CREATORS_COUNT = 95
 CREATOR_ID = 40439
 CREATOR_NAME = "Geoff Johns"
 CREATOR_COUNT = "67"
@@ -11,7 +10,6 @@ FIRST_ISSUE_NAME = "Airborne"
 FIRST_ISSUE_NUMBER = "1"
 ID = 18216
 ISSUE_COUNT = 67
-ISSUES_COUNT = 67
 ISSUE_ID = 106713
 ISSUE_NAME = "Wanted: Hal Jordan Part 4"
 ISSUE_NUMBER = "17"
@@ -26,7 +24,6 @@ START_YEAR = "2005"
 
 def test_volume(talker):
     result = talker.volume(ID)
-    assert len(result.creators) == CREATORS_COUNT
     assert result.creators[0].id == CREATOR_ID
     assert result.creators[0].name == CREATOR_NAME
     assert result.creators[0].count == CREATOR_COUNT
@@ -35,7 +32,6 @@ def test_volume(talker):
     assert result.first_issue.number == FIRST_ISSUE_NUMBER
     assert result.id == ID
     assert result.issue_count == ISSUE_COUNT
-    assert len(result.issues) == ISSUES_COUNT
     assert result.issues[0].id == ISSUE_ID
     assert result.issues[0].name == ISSUE_NAME
     assert result.issues[0].number == ISSUE_NUMBER
@@ -56,13 +52,11 @@ def test_volume_fail(talker):
 def test_volume_list(talker):
     search_results = talker.volume_list({"filter": f"name:{NAME}"})
     result = [x for x in search_results if x.id == ID][0]
-    # Search doesn't contain Creators
     assert result.first_issue.id == FIRST_ISSUE_ID
     assert result.first_issue.name == FIRST_ISSUE_NAME
     assert result.first_issue.number == FIRST_ISSUE_NUMBER
     assert result.id == ID
     assert result.issue_count == ISSUE_COUNT
-    # Search doesn't contain Issues
     assert result.last_issue.id == LAST_ISSUE_ID
     assert result.last_issue.name == LAST_ISSUE_NAME
     assert result.last_issue.number == LAST_ISSUE_NUMBER
